feat: add rest block activity
