+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,6 @@
 
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE_DATA_SYNC" />
-    <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW" />
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
     <uses-permission android:name="android.permission.PACKAGE_USAGE_STATS"
         xmlns:tools="http://schemas.android.com/tools"
@@ -16,6 +15,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:theme="@style/Theme.ScreenCycle">
 
+        <activity android:name=".ui.BlockActivity" />
         <activity android:name=".ui.AppSelectionActivity" />
         <activity
             android:name=".ui.MainActivity"
@@ -30,10 +30,6 @@
             android:name=".core.CycleService"
             android:exported="false"
             android:foregroundServiceType="dataSync" />
-
-        <service
-            android:name=".core.BlockOverlayService"
-            android:exported="false" />
 
         <service
             android:name=".core.AppAccessibilityService"
