+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="accessibility_description">Track active app to block games during REST</string>
+    <string name="rest_now">Сейчас отдых</string>
 </resources>
