+++ docs/ui-guide.md
@@ -22,7 +22,7 @@
 
 Приложение использует тему Material You и автоматически переключает светлый и тёмный режим в зависимости от настроек системы.
 
-Во время отдыха приложение показывает полноэкранный оверлей, блокирующий ввод. Оверлей работает на Android 8+ без дополнительных условий.
+Во время отдыха приложение возвращает пользователя на главный экран и показывает полноэкранное окно «Сейчас отдых». Окно работает на Android 8+ без дополнительных условий.
 
 ## Таймер цикла / Cycle timer (с версии Unreleased)
